+++ docs/news.rst
@@ -7,6 +7,14 @@
 ``psycopg`` release notes
 =========================
 
+Psycopg 3.2.10 (unreleased)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+- Fix memory leak when lambda/local functions are used as argument for
+  `~.psycopg.types.json.set_json_dumps()`, `~.psycopg.types.json.set_json_loads()`
+  (:ticket:`#1108`).
+
+
 Current release
 ---------------